+++ src/features/game/WatchPointTab.tsx
@@ -1,0 +1,5 @@
+const WatchPointTab = () => {
+  return <div>WatchPointTab</div>;
+};
+
+export { WatchPointTab };
