+++ src/features/game/index.tsx
@@ -1,0 +1,3 @@
+export { BoxScoreTab } from './BoxScoreTab';
+export { MatchScheduleTab } from './MatchScheduleTab';
+export { WatchPointTab } from './WatchPointTab';
